Fix loading package.json to find version
In `lib/kernel.js`, line 183 tries to load ijavascript's `package.json` file to find the current version number. However, if $PWD isn't the ijavascript module folder, this will fail with a "file not found" error.

This commit changes that line so that it tries to load `__dirname/../package.json` (using the `path.join()` utility for cross-platform portability).